refactor(comp:i18n): update locales
