--- packages/components/i18n/src/locales/index.ts
@@ -1,9 +1,0 @@
-/**
- * @license
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://github.com/IDuxFE/idux/blob/main/LICENSE
- */
-
-export * from './en-US'
-export * from './zh-CN'
