+++ packages/site/src/docs/I18n.zh.md
@@ -9,18 +9,18 @@ order: 4
 ## 初始化
 
 ```typescript
-import { useLocale, en_US } from '@idux/components/i18n'
+import { useLocale, enUS } from '@idux/components/i18n'
 
-useLocale(en_US)
+useLocale(enUS)
 ```
 
 ## 运行时切换
 
 ```typescript
-import { addLocale, useLocale, zh_CN, en_US } from '@idux/components/i18n'
+import { addLocale, useLocale, zhCN, enUS } from '@idux/components/i18n'
 
 // 首先需要先添加多语言包
-addLocale([zh_CN, en_US])
+addLocale([zhCN, enUS])
 useLocale('en-US')
 // 运行时切换
 setTimeout(()=> useLocale('zh-CN'), 3000)
